Move 2D benchmarks into their own directory/crate.
