+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = [ "build/rapier2d", "build/rapier_testbed2d", "examples2d",
+members = [ "build/rapier2d", "build/rapier_testbed2d", "examples2d", "benchmarks2d",
             "build/rapier3d", "build/rapier_testbed3d", "examples3d", "benchmarks3d" ]
 
 [patch.crates-io]
