+++ benchmarks2d/boxes2.rs
